Hitboxer expanded and Input class remade
Hitboxer objects now have support for death and death animation.

Input class is now it's own class again and  saves inputs in an list and
emties the list every tick when Game class retrieves the list.

Input reactions are back in the Engine loop.

Resourceloader is moved to the Engine because more rappid requests to
that class.

Also deleted out commented code that isn't needed